improvement/OSIS-3810 move organization_url to Campus
