+++ learning_unit/api/views/learning_unit.py
@@ -82,7 +82,7 @@ class LearningUnitDetailed(LanguageContextSerializerMixin, generics.RetrieveAPIV
         ).prefetch_related(
             'learning_container_year__requirement_entity__entityversion_set',
             'learningcomponentyear_set',
-        ).annotate_campus_website().annotate_full_title()
+        ).annotate_full_title()
         luy = get_object_or_404(
             LearningUnitYearQuerySet.annotate_entities_allocation_and_requirement_acronym(queryset),
             acronym__iexact=acronym,
